fix(jobs, events): add correct imports and add file in the correct  folders for speration of concerns
