+++ package.json
@@ -104,12 +104,13 @@
     "moduleNameMapper": {
       "@emails/(.*)$": "<rootDir>/emails/$1",
       "@notification/(.*)$": "<rootDir>/notification/$1",
-      "@events/(.*)$": "<rootDir>/events/$1",
+      "@jobs/(.*)$": "<rootDir>/jobs/$1",
       "@reports/(.*)$": "<rootDir>/reports/$1",
       "@common/(.*)$": "<rootDir>/common/$1",
       "@shared/(.*)$": "<rootDir>/shared/$1",
       "@databases/(.*)$": "<rootDir>/databases/$1",
       "@configuration/(.*)$": "<rootDir>/configuration/$1",
+      "@events/(.*)$": "<rootDir>/core/events/$1",
       "@accesses/(.*)$": "<rootDir>/core/accesses/$1",
       "@admins/(.*)$": "<rootDir>/core/admins/$1",
       "@auth/(.*)$": "<rootDir>/core/auth/$1",
