+++ src/core/events/entities/event.entity.ts
