+++ src/core/events/events.service.spec.ts
@@ -3,7 +3,7 @@ import { EventEmitter2 } from '@nestjs/event-emitter';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { EventsService } from './events.service';
 import { Event } from './entities/event.entity';
-import { ScheduleService } from './schedule.service';
+import { ScheduleService } from '@jobs/schedule.service';
 import { CreateEventDto } from './dto/create-event.dto';
 import { UpdateEventDto } from './dto/update-event.dto';
 import { JobInfo } from '@common/interfaces/schedule';
