+++ src/core/events/interfaces/event.ts
