+++ src/jobs/interfaces/jobs.ts
