+++ src/jobs/jobs.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
-import { JobsService } from './jobs.service';
-import { JobsController } from './jobs.controller';
+import { JobsService } from '@jobs/jobs.service';
+import { JobsController } from '@jobs/jobs.controller';
+import { ScheduleService } from '@jobs/schedule.service';
 
 @Module({
   controllers: [JobsController],
-  providers: [JobsService],
+  providers: [JobsService, ScheduleService],
+  exports: [JobsService, ScheduleService],
 })
 export class JobsModule {}
